+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 
 ### Version 11
 
+- Show number of different solutions found by the AI
 - cycle through different solutions if more than one solution is found
 
 ### Version 10.1